move db files - migrations, seeders, data/gen to db folder
