+++ .gitignore
@@ -49,7 +49,7 @@ Thumbs.db
 /.sass-cache/
 
 /css/***
-/dist/***
+/dist/**/*
 
 # environment/config #
 ######################
