+++ .sequelizerc
@@ -3,14 +3,14 @@ const path = require('path');
 if (process.env !== 'production') {
   module.exports = {
     config: path.resolve('./config', 'config.js'),
-    //  "models-path": path.resolve('./models'),
-    //  "seeders-path": path.resolve('./seeders'),
-    //  "migrations-path": path.resolve('./migrations')
+     "models-path": path.resolve('./dist/models'),
+     "seeders-path": path.resolve('./db/seeders'),
+     "migrations-path": path.resolve('./db/migrations')
   };
 }
 module.exports = {
   config: path.resolve('./config', 'config.js'),
-  // "models-path": path.resolve('./server/dist/models'),
-  // "seeders-path": path.resolve('./server/dist/seeders'),
-  // "migrations-path": path.resolve('./server/dist/migrations')
+  "models-path": path.resolve('./dist/models'),
+  "seeders-path": path.resolve('./db/seeders'),
+  "migrations-path": path.resolve('./db/migrations')
 };
