+++ db/data/generation/generate-json-data.js
@@ -5,7 +5,6 @@ require('./users');
 require('./venues');
 require('./events');
 require('./bands');
-// join tables
 require('./venue-genres');
 require('./users-bands');
 require('./users-venues');
