--- src/data/event-bands.json
@@ -1,1 +1,0 @@
-[{"EventId":1,"BandId":57,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-06-30T04:19:04.783Z"},{"EventId":1,"BandId":12,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-05-20T11:28:36.648Z"},{"EventId":2,"BandId":25,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-02-08T20:28:43.582Z"},{"EventId":2,"BandId":19,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-06-05T19:08:55.940Z"},{"EventId":2,"BandId":16,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-03-23T10:46:24.935Z"},{"EventId":2,"BandId":12,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-07-07T00:32:25.950Z"},{"EventId":2,"BandId":29,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-08-16T02:58:01.837Z"},{"EventId":3,"BandId":14,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-07-28T16:06:08.098Z"},{"EventId":3,"BandId":52,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-12-20T02:30:56.288Z"},{"EventId":3,"BandId":4,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-07-23T16:13:03.769Z"},{"EventId":4,"BandId":12,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-10-04T15:07:13.916Z"},{"EventId":4,"BandId":50,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-02-09T16:45:46.790Z"},{"EventId":4,"BandId":44,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-02-22T07:55:58.226Z"},{"EventId":4,"BandId":33,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-06-25T08:11:56.264Z"},{"EventId":4,"BandId":25,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-09-23T17:18:38.208Z"},{"EventId":4,"BandId":45,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-10-06T00:17:43.981Z"},{"EventId":5,"BandId":52,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-04-30T15:42:57.018Z"},{"EventId":6,"BandId":43,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-01-20T09:11:07.174Z"},{"EventId":6,"BandId":13,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-05-14T15:30:09.803Z"},{"EventId":6,"BandId":48,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-08-31T16:27:58.162Z"},{"EventId":7,"BandId":57,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-11-20T16:30:52.173Z"},{"EventId":7,"BandId":18,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-03-18T00:28:53.593Z"},{"EventId":7,"BandId":9,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-03-03T09:08:53.139Z"},{"EventId":7,"BandId":36,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-03-05T23:20:37.214Z"},{"EventId":7,"BandId":44,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-08-02T20:30:52.132Z"},{"EventId":8,"BandId":54,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-01-27T16:33:12.401Z"},{"EventId":9,"BandId":18,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-10-23T13:00:49.512Z"},{"EventId":9,"BandId":27,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-03-08T21:42:02.446Z"},{"EventId":10,"BandId":30,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-04-05T02:22:43.923Z"},{"EventId":10,"BandId":57,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-02-01T07:46:45.679Z"},{"EventId":10,"BandId":47,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-06-29T18:34:42.998Z"},{"EventId":11,"BandId":46,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-08-06T19:38:38.984Z"},{"EventId":11,"BandId":23,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-11-14T17:32:06.622Z"},{"EventId":12,"BandId":44,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-12-17T13:47:25.544Z"},{"EventId":12,"BandId":59,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-08-20T19:15:30.671Z"},{"EventId":12,"BandId":56,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-11-30T01:22:09.411Z"},{"EventId":12,"BandId":36,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-03-12T19:40:45.737Z"},{"EventId":12,"BandId":52,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-06-13T18:12:13.224Z"},{"EventId":13,"BandId":51,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-10-11T01:51:20.477Z"},{"EventId":13,"BandId":19,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-02-03T22:02:45.465Z"},{"EventId":13,"BandId":37,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-08-12T22:09:57.901Z"},{"EventId":13,"BandId":42,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-02-14T08:37:06.676Z"},{"EventId":13,"BandId":33,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-07-30T05:18:28.798Z"},{"EventId":14,"BandId":8,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-09-07T06:40:15.313Z"},{"EventId":15,"BandId":37,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-03-09T14:00:49.428Z"},{"EventId":15,"BandId":30,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-06-05T04:09:37.692Z"},{"EventId":15,"BandId":59,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-11-05T23:31:59.632Z"},{"EventId":16,"BandId":34,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-09-14T07:17:34.632Z"},{"EventId":16,"BandId":42,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-02-18T14:59:32.354Z"},{"EventId":17,"BandId":57,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-05-18T10:25:08.496Z"},{"EventId":17,"BandId":13,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-08-17T23:25:05.291Z"},{"EventId":17,"BandId":33,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-01-17T08:44:22.836Z"},{"EventId":18,"BandId":33,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2017-08-29T14:11:01.713Z"},{"EventId":18,"BandId":24,"updatedAt":"2018-06-03T20:43:09.747Z","createdAt":"2018-05-31T12:30:14.960Z"},{"EventId":18,"BandId":13,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2017-11-02T08:02:45.428Z"},{"EventId":18,"BandId":60,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2017-12-13T07:17:45.636Z"},{"EventId":19,"BandId":56,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2017-08-16T02:11:58.816Z"},{"EventId":19,"BandId":55,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2018-04-23T04:55:46.689Z"},{"EventId":19,"BandId":13,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2017-07-22T04:17:50.093Z"},{"EventId":19,"BandId":15,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2017-10-26T06:51:43.090Z"},{"EventId":20,"BandId":7,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2017-07-27T11:35:19.733Z"},{"EventId":20,"BandId":23,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2018-01-11T10:32:37.961Z"},{"EventId":20,"BandId":54,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2017-12-30T00:25:13.683Z"},{"EventId":20,"BandId":55,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2017-09-11T06:50:40.163Z"},{"EventId":20,"BandId":10,"updatedAt":"2018-06-03T20:43:09.748Z","createdAt":"2018-05-30T10:06:30.633Z"}]
