--- src/data/events.json
@@ -1,1 +1,0 @@
-[{"VenueId":10,"startsAt":"2018-07-05T13:32:34.171Z","endsAt":"2018-07-05T16:32:34.171Z","title":"Handcrafted Wooden Car","description":"Dolorum voluptas quis.","updatedAt":"2018-06-03T20:43:09.732Z","createdAt":"2017-10-13T04:17:34.760Z"},{"VenueId":7,"startsAt":"2018-06-30T21:09:06.868Z","endsAt":"2018-07-01T00:09:06.868Z","title":"Awesome Cotton Pants","description":"Omnis qui quis itaque sit.","updatedAt":"2018-06-03T20:43:09.733Z","createdAt":"2017-07-04T10:05:06.237Z"},{"VenueId":7,"startsAt":"2017-11-30T09:27:28.099Z","endsAt":"2017-11-30T12:27:28.099Z","title":"Awesome Concrete Keyboard","description":"Corrupti ab quo consequatur.","updatedAt":"2018-06-03T20:43:09.733Z","createdAt":"2017-12-29T06:56:45.332Z"},{"VenueId":1,"startsAt":"2018-07-09T19:12:57.725Z","endsAt":"2018-07-09T22:12:57.725Z","title":"Ergonomic Frozen Ball","description":"Et tenetur magni est adipisci.","updatedAt":"2018-06-03T20:43:09.733Z","createdAt":"2017-10-20T02:03:38.730Z"},{"VenueId":7,"startsAt":"2018-08-14T00:33:34.559Z","endsAt":"2018-08-14T03:33:34.559Z","title":"Rustic Soft Pants","description":"Odio impedit magni consequuntur eaque expedita maxime sint.","updatedAt":"2018-06-03T20:43:09.733Z","createdAt":"2017-12-05T18:58:02.571Z"},{"VenueId":2,"startsAt":"2018-01-05T11:12:01.759Z","endsAt":"2018-01-05T14:12:01.759Z","title":"Intelligent Frozen Soap","description":"Molestias repudiandae aliquid molestiae neque earum sit.","updatedAt":"2018-06-03T20:43:09.733Z","createdAt":"2017-08-17T09:52:44.503Z"},{"VenueId":9,"startsAt":"2017-10-23T09:32:29.435Z","endsAt":"2017-10-23T12:32:29.435Z","title":"Rustic Granite Shirt","description":"Omnis quas neque iste esse eos sit quisquam quo.","updatedAt":"2018-06-03T20:43:09.734Z","createdAt":"2017-09-16T08:19:26.768Z"},{"VenueId":6,"startsAt":"2017-11-17T14:30:14.448Z","endsAt":"2017-11-17T17:30:14.448Z","title":"Gorgeous Fresh Towels","description":"Deserunt illum aut iure consequatur.","updatedAt":"2018-06-03T20:43:09.734Z","createdAt":"2017-06-19T00:15:22.232Z"},{"VenueId":8,"startsAt":"2018-05-19T10:03:27.112Z","endsAt":"2018-05-19T13:03:27.112Z","title":"Fantastic Wooden Shirt","description":"Recusandae placeat assumenda natus.","updatedAt":"2018-06-03T20:43:09.734Z","createdAt":"2017-06-06T11:16:15.839Z"},{"VenueId":9,"startsAt":"2019-03-22T14:00:50.607Z","endsAt":"2019-03-22T17:00:50.607Z","title":"Generic Frozen Ball","description":"Reiciendis cupiditate in ut omnis saepe sint numquam nostrum.","updatedAt":"2018-06-03T20:43:09.734Z","createdAt":"2017-10-09T21:47:14.844Z"},{"VenueId":9,"startsAt":"2018-01-19T10:16:51.235Z","endsAt":"2018-01-19T13:16:51.235Z","title":"Tasty Metal Computer","description":"Perspiciatis nobis facere omnis asperiores sed.","updatedAt":"2018-06-03T20:43:09.734Z","createdAt":"2018-01-13T21:21:21.852Z"},{"VenueId":7,"startsAt":"2018-07-24T11:17:26.472Z","endsAt":"2018-07-24T14:17:26.472Z","title":"Refined Plastic Chips","description":"Tempora officiis autem illo pariatur repellendus ducimus voluptas.","updatedAt":"2018-06-03T20:43:09.734Z","createdAt":"2018-03-02T23:09:51.491Z"},{"VenueId":3,"startsAt":"2018-06-06T18:37:43.379Z","endsAt":"2018-06-06T21:37:43.379Z","title":"Handcrafted Wooden Bacon","description":"Velit autem accusamus quasi ducimus illum molestiae a amet id.","updatedAt":"2018-06-03T20:43:09.734Z","createdAt":"2017-09-10T04:50:14.723Z"},{"VenueId":8,"startsAt":"2018-07-08T20:25:47.173Z","endsAt":"2018-07-08T23:25:47.173Z","title":"Gorgeous Cotton Keyboard","description":"Ipsum ex qui rerum aliquid nemo.","updatedAt":"2018-06-03T20:43:09.735Z","createdAt":"2017-12-07T09:38:12.513Z"},{"VenueId":9,"startsAt":"2018-05-24T10:53:49.299Z","endsAt":"2018-05-24T13:53:49.299Z","title":"Handcrafted Fresh Fish","description":"Ipsum et et omnis dolores ad voluptatem perferendis occaecati.","updatedAt":"2018-06-03T20:43:09.735Z","createdAt":"2017-12-03T05:36:14.449Z"},{"VenueId":1,"startsAt":"2018-12-07T10:58:27.116Z","endsAt":"2018-12-07T13:58:27.116Z","title":"Unbranded Rubber Ball","description":"Et id quos et tempora ut.","updatedAt":"2018-06-03T20:43:09.735Z","createdAt":"2018-03-15T13:48:27.515Z"},{"VenueId":6,"startsAt":"2018-01-17T06:13:51.912Z","endsAt":"2018-01-17T09:13:51.912Z","title":"Small Granite Shirt","description":"Eligendi dolore aut ab necessitatibus est veniam consequuntur illum.","updatedAt":"2018-06-03T20:43:09.735Z","createdAt":"2017-09-22T05:03:59.779Z"},{"VenueId":1,"startsAt":"2018-02-14T00:13:04.224Z","endsAt":"2018-02-14T03:13:04.224Z","title":"Licensed Wooden Mouse","description":"Sit illo aut molestias qui aut accusantium provident voluptatum.","updatedAt":"2018-06-03T20:43:09.735Z","createdAt":"2017-09-30T08:24:55.452Z"},{"VenueId":7,"startsAt":"2018-04-17T17:29:35.842Z","endsAt":"2018-04-17T20:29:35.842Z","title":"Licensed Fresh Chicken","description":"Aperiam voluptatem excepturi ut quis repudiandae nisi.","updatedAt":"2018-06-03T20:43:09.736Z","createdAt":"2018-01-11T00:13:11.944Z"},{"VenueId":8,"startsAt":"2018-11-02T17:47:02.494Z","endsAt":"2018-11-02T20:47:02.494Z","title":"Refined Frozen Chair","description":"Vero dolorem qui cum ut omnis consequatur ipsam impedit.","updatedAt":"2018-06-03T20:43:09.736Z","createdAt":"2017-09-04T18:20:53.697Z"}]
