--- src/data/generation/generator-amounts.json
@@ -1,17 +1,0 @@
-{
-	"amounts": {
-    "numUsers": 200,
-		"numBands": 60,
-		"numVenues": 10,
-		"numEvents": 20,
-		"numGenres": 5,
-    "numAddresses": 250,
-    "numEventBands": 100,
-    "numUsersBands": 150,
-    "numUsersVenues": 30,
-    "numVenueGenres": 20,
-    "maxBandsPerEvent": 6,
-    "maxUsersPerBand": 6,
-    "maxUsersPerVenue": 3
-	}
-}
