--- src/data/genres.json
@@ -1,1 +1,0 @@
-[{"name":"Blues","updatedAt":"2018-06-03T20:42:55.416Z","createdAt":"2018-05-21T23:08:47.127Z"},{"name":"Country","updatedAt":"2018-06-03T20:42:55.416Z","createdAt":"2018-05-29T20:39:47.235Z"},{"name":"Electronic","updatedAt":"2018-06-03T20:42:55.416Z","createdAt":"2018-06-02T18:19:40.039Z"},{"name":"Hip-Hop","updatedAt":"2018-06-03T20:42:55.416Z","createdAt":"2017-11-05T19:44:06.658Z"},{"name":"Rock","updatedAt":"2018-06-03T20:42:55.416Z","createdAt":"2017-12-13T23:31:59.117Z"}]
