--- src/data/users-venues.json
@@ -1,1 +1,0 @@
-[{"VenueId":1,"UserId":64,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2017-12-24T18:38:40.405Z"},{"VenueId":2,"UserId":49,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2018-03-17T23:53:03.437Z"},{"VenueId":3,"UserId":197,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2017-08-24T18:56:15.414Z"},{"VenueId":3,"UserId":73,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2018-05-23T14:30:39.331Z"},{"VenueId":4,"UserId":172,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2017-08-18T02:10:00.972Z"},{"VenueId":4,"UserId":7,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2018-06-03T13:30:12.086Z"},{"VenueId":4,"UserId":39,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2017-06-19T22:10:58.750Z"},{"VenueId":5,"UserId":167,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2018-04-05T02:53:36.880Z"},{"VenueId":6,"UserId":163,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2017-10-21T16:07:22.622Z"},{"VenueId":6,"UserId":56,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2018-04-21T17:40:29.199Z"},{"VenueId":6,"UserId":36,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2017-10-10T18:51:59.219Z"},{"VenueId":7,"UserId":104,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2018-05-23T14:50:55.734Z"},{"VenueId":7,"UserId":21,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2017-12-12T02:40:20.455Z"},{"VenueId":7,"UserId":120,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2017-09-07T11:00:28.614Z"},{"VenueId":8,"UserId":17,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2017-06-13T04:46:24.233Z"},{"VenueId":9,"UserId":35,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2018-04-21T03:50:08.050Z"},{"VenueId":9,"UserId":25,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2018-01-29T15:27:29.473Z"},{"VenueId":9,"UserId":64,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2018-04-12T18:29:05.545Z"},{"VenueId":10,"UserId":152,"updatedAt":"2018-06-03T20:43:09.746Z","createdAt":"2018-04-18T15:34:36.701Z"}]
