--- src/data/venue-genres.json
@@ -1,1 +1,0 @@
-[{"VenueId":1,"GenreId":3,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-01-20T13:43:23.749Z"},{"VenueId":1,"GenreId":4,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-04-10T18:00:11.469Z"},{"VenueId":2,"GenreId":5,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2017-12-17T17:05:35.972Z"},{"VenueId":2,"GenreId":2,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2017-06-08T16:35:10.899Z"},{"VenueId":2,"GenreId":4,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-03-01T06:10:28.591Z"},{"VenueId":3,"GenreId":2,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-05-06T07:41:19.081Z"},{"VenueId":3,"GenreId":3,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-05-17T09:52:52.755Z"},{"VenueId":4,"GenreId":3,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2017-06-23T21:45:06.450Z"},{"VenueId":4,"GenreId":2,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2017-07-26T10:54:23.184Z"},{"VenueId":4,"GenreId":1,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-01-11T04:04:03.835Z"},{"VenueId":4,"GenreId":5,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-03-05T18:45:50.394Z"},{"VenueId":5,"GenreId":5,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-05-21T06:39:41.177Z"},{"VenueId":5,"GenreId":2,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2017-10-21T12:04:20.534Z"},{"VenueId":6,"GenreId":3,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2017-12-29T23:50:06.954Z"},{"VenueId":7,"GenreId":2,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-01-18T09:14:28.342Z"},{"VenueId":7,"GenreId":1,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-02-28T16:20:11.304Z"},{"VenueId":7,"GenreId":3,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-03-05T03:38:11.920Z"},{"VenueId":8,"GenreId":5,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2017-12-28T13:20:05.401Z"},{"VenueId":9,"GenreId":5,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2017-08-28T03:01:38.235Z"},{"VenueId":9,"GenreId":1,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2017-11-18T06:00:41.938Z"},{"VenueId":9,"GenreId":4,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-01-25T21:24:22.116Z"},{"VenueId":10,"GenreId":2,"updatedAt":"2018-06-03T20:43:09.743Z","createdAt":"2018-02-24T22:21:20.667Z"}]
