--- src/data/venues.json
@@ -1,1 +1,0 @@
-[{"name":"Corkery","AddressId":221,"updatedAt":"2018-06-03T20:43:09.716Z","createdAt":"2018-02-07T11:06:53.522Z"},{"name":"Swift","AddressId":3,"updatedAt":"2018-06-03T20:43:09.716Z","createdAt":"2017-09-14T22:45:35.756Z"},{"name":"Abshire","AddressId":109,"updatedAt":"2018-06-03T20:43:09.716Z","createdAt":"2018-05-26T22:14:39.051Z"},{"name":"Abbott","AddressId":250,"updatedAt":"2018-06-03T20:43:09.716Z","createdAt":"2018-05-26T13:19:13.719Z"},{"name":"Keebler","AddressId":78,"updatedAt":"2018-06-03T20:43:09.716Z","createdAt":"2017-09-13T09:34:18.253Z"},{"name":"Cronin","AddressId":110,"updatedAt":"2018-06-03T20:43:09.716Z","createdAt":"2018-04-27T14:35:03.593Z"},{"name":"Reichel","AddressId":65,"updatedAt":"2018-06-03T20:43:09.716Z","createdAt":"2017-07-27T21:58:35.630Z"},{"name":"Beatty","AddressId":236,"updatedAt":"2018-06-03T20:43:09.716Z","createdAt":"2018-01-15T00:10:47.781Z"},{"name":"Raynor","AddressId":221,"updatedAt":"2018-06-03T20:43:09.716Z","createdAt":"2017-11-21T06:59:44.608Z"},{"name":"Bechtelar","AddressId":158,"updatedAt":"2018-06-03T20:43:09.716Z","createdAt":"2017-12-14T00:22:51.117Z"}]
